+++ db/migrate/20230804140435_add_last_session_column_to_users.rb
@@ -1,0 +1,5 @@
+class AddLastSessionColumnToUsers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :latest_session, :datetime
+  end
+end